Reorganize dataset tests
